+++ src/modules/avatar/avatar.types.ts
@@ -34,13 +34,19 @@ export interface IProps {
   markerOffsetRight?: number;
 
   /** Size of the avatar. */
-  size?: number;
+  size?: 'tile' | 'profile' | number;
 
   /** Position of marker from bottom. */
   markerOffsetBottom?: number;
 
   /** Renders a white icon instead of a black. */
   inversed?: boolean;
+
+  /** Adds border radius of the specified size. */
+  radius?: 'none' | 'small' | 'full';
+
+  /** Shows the initials as placeholder when image is unavailable. */
+  initials?: string;
 
   /** Function to be called when the avatar is pressed. */
   onPress?: any;